trading-bridge/requirements.txt
pyzmq>=25.1.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
schedule>=1.2.0
pywin32>=306; sys_platform == 'win32'

# AI/ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pandas-ta>=0.3.14b
yfinance>=0.2.28
ccxt>=4.0.0

# Deep Learning (choose one)
tensorflow>=2.13.0
# torch>=2.0.0  # Alternative: PyTorch

# Technical Analysis (optional - requires TA-Lib C library)
# ta-lib>=0.4.28

